+++ contrib/deneva/concurrency_control/unified_util.h
@@ -11,13 +11,13 @@
 #ifndef _ROW_UNIFIED_UTIL_H_
 #define _ROW_UNIFIED_UTIL_H_
 
-#include "../unified_concurrency_control/row_prece.h"
-#include "../unified_concurrency_control/row_ssi.h"
-#include "../unified_concurrency_control/alg_dli_identify_cycle.h"
-#include "../unified_concurrency_control/alg_dli_identify_chain.h"
-#include "../unified_concurrency_control/alg_ssi.h"
-#include "../unified_concurrency_control/txn_dli_identify.h"
-#include "../unified_concurrency_control/txn_ssi.h"
+#include "../../../src/3ts/backend/cca/unified_history_algorithm/row/row_prece.h"
+#include "../../../src/3ts/backend/cca/unified_history_algorithm/row/row_ssi.h"
+#include "../../../src/3ts/backend/cca/unified_history_algorithm/alg/alg_dli_identify_cycle.h"
+#include "../../../src/3ts/backend/cca/unified_history_algorithm/alg/alg_dli_identify_chain.h"
+#include "../../../src/3ts/backend/cca/unified_history_algorithm/alg/alg_ssi.h"
+#include "../../../src/3ts/backend/cca/unified_history_algorithm/txn/txn_dli_identify.h"
+#include "../../../src/3ts/backend/cca/unified_history_algorithm/txn/txn_ssi.h"
 
 class row_t;
