+++ src/3ts/backend/cca/unified_history_algorithm/alg/alg_ssi.h
@@ -11,14 +11,14 @@
 
 #pragma once
 
-#include "dli_identify_util.h"
-#include "row_prece.h"
+#include "../util/dli_identify_util.h"
+//#include "../row/row_prece.h"
 #include <mutex>
 #include <type_traits>
 #include <list>
 #include <algorithm>
 #include <deque>
-#include "row_ssi.h"
+#include "../row/row_ssi.h"
 
 namespace ttts {
