+++ src/3ts/backend/cca/unified_history_algorithm/row/row_ssi.h
@@ -10,7 +10,7 @@
 
 #pragma once
 
-#include "util.h"
+#include "../util/util.h"
 #include <type_traits>
 #include <deque>
 #include <memory>
