+++ src/3ts/backend/cca/unified_history_algorithm/unified_history_algorithm.h
@@ -1,16 +1,16 @@
-#include "algorithm.h"
-#include "../../../../contrib/deneva/unified_concurrency_control/util.h"
-#include "anomaly_type.h"
-#include "prece_type.h"
-#include "algorithm.h"
-#include "../../../../contrib/deneva/unified_concurrency_control/dli_identify_util.h"
-#include "../../../../contrib/deneva/unified_concurrency_control/txn_dli_identify.h"
-#include "../../../../contrib/deneva/unified_concurrency_control/txn_ssi.h"
-#include "../../../../contrib/deneva/unified_concurrency_control/row_prece.h"
-#include "../../../../contrib/deneva/unified_concurrency_control/row_ssi.h"
-#include "../../../../contrib/deneva/unified_concurrency_control/alg_dli_identify_chain.h"
-#include "../../../../contrib/deneva/unified_concurrency_control/alg_dli_identify_cycle.h"
-#include "../../../../contrib/deneva/unified_concurrency_control/alg_ssi.h"
+#include "../algorithm.h"
+#include "util/util.h"
+#include "util/anomaly_type.h"
+#include "util/prece_type.h"
+#include "../algorithm.h"
+#include "util/dli_identify_util.h"
+#include "txn/txn_dli_identify.h"
+#include "txn/txn_ssi.h"
+#include "row/row_prece.h"
+#include "row/row_ssi.h"
+#include "alg/alg_dli_identify_chain.h"
+#include "alg/alg_dli_identify_cycle.h"
+#include "alg/alg_ssi.h"
 #include <optional>
 
 namespace ttts {
